// src/endpoints/domain/dtos/profile.dto.ts
export default class ProfileDto {
  name?: string;
  username?: string;
  avatar?: string;
  location?: string;
  website?: string;
  shortbio?: string;
  telegram?: string;
  medium?: string;
  discord?: string;
  twitter?: string;
  facebook?: string;
  otherLink?: string;
  textRecords?: object[];
  walletEgld?: string;
  walletEth?: string;
  walletBtc?: string;
}